Run update jobs for fuller set of services
Normally the listed services are omitted from the scenario jobs
because they are often covered by many other "default" jobs. However,
in the case of the update job, this is our only update job, so we need
to try to run it on a full set of services which are deployed by the
job. Otherwise we risk breaking the gate by merging something that
affects update job without testing it by the update job.

Change-Id: I8115333bb3be27a461f7dab9aba20749ed695a49