Remove legacy controller-v6.yaml NIC config templates
Recent changes to the network interface configurations in THT
make it unneccessary to use the special controller-v6.yaml
template. Instead, adding both an IPv6 and an IPv4 network
to the default_route_networks: parameter in roles_data.yaml
will result in both an IPv4 and an IPv6 default route.

For instance:
  default_route_networks: ['ControlPlane', 'External']

Change-Id: I23ae304362bd0029cc921010155414f023b19c54
Closes-bug: 1793581