Add network data for use with openshift deployments
Provide a network data file with only the External, InternalApi and the
Storage networks for use with OpenShift deployments.

Take out the IPv6 addresses from the network data until OpenShift can
be deployed in an IPv6 environment.

This also disables the validation for the network_data_openshift.yaml
since it expects network_data files to only add new networks to the
existing pre-defined overcloud networks and we're doing the opposite
here -- only keeping a subset of the networks.

Change-Id: I6d8c08c12ae9002e6386f26c5a8bcf70b8eda4b7