Ensure directories managed by libvirt-daemon exist
The /var/run/libvirt directory is managed by libvirt-daemon. As
this service is now running within a container, the directory
might not exists on a fresh install.
This patch also takes care of the fact the /var/run directory is
cleaned upon reboot, and uses systemd-tmpfiles in order to ensure
the directory creation during the boot process.

We also take care of missing log directories, also managed by the
libvirt-daemon package.

Change-Id: I6f8dc49ff556215a25e8bb23ae8da63ce607d70c
Closes-Bug: #1801977