Do not ignore Swift ring changes to trigger container restart
Swift containers need to restart if the rings change. In
non-containerized environments this is not required, because the Swift
processes will reload the rings on any changes. However, this does not
work within containers, thus a restart is required.

This also restarts swift_copy_rings and swift_setup_srv container. This
will copy the updated ring files and ensure new storage mount points are
using the right permissions.

Closes-Bug: 1802066
Related-Bug: 1786065
Change-Id: Ie2b9f003dc34f2f02a45293d06d6a40c8d5ed8ff