Switch stable scenarios to voting
In master we have removed these from the gate as they are non-voting
because we plan on replacing them with standalone versions. However for
stable branches we will want to leave these as voting since they are
defined in the gate in the branch layout.

Change-Id: I7a859ac1db9206301df897553320fca70ea33980
Related-Bug: #1802144
(cherry picked from commit 3061e5d1bce513b3d58eae963722569b7f9bb50b)