Add OVN environment that works with standalone mode.
All of the current OVN environment files enable HA for the OVN
database, which does not work with tripleo standalone.  This
introduces an environment that can be used in that mode.

There used to be non-HA modes, but they were removed in commit
819805d708cccc6e1419a47b16fc012955ce9156.  This adds back just one,
and gives it a "standalone" name to help clarify that this is not
one of the scenarios intended for full production deployments.

Change-Id: Ie21c468d1cf7e4db9c406c62f5d09f6af97d593a