Use ipc: host with containers that use pcs
We currently bind-mount /dev/shm inside the containers that need
to run pcs commands (like in the temporary containers that need to create
pcmk resources via puppet). The reason for this is the way CLI commands
interact with the cluster: they need access to /dev/shm for them to
work.

With docker we used to simply bind mount /dev/shm from the host, but
this does not work with podman because bind mounts under /dev seem to
have no effect.

The correct way to do this is to use '--ipc=host' which will do the
right thing for both podman and docker.

Tested on a docker backend by doing a fresh 3Ctrl + 2Cmp deploy and
then a redeploy to verify that restart bundles are working.

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>

Change-Id: Id0e737af4cb82d56b0c2db3303a767bb671917ea