Configure registry to use gluster
We're deploying an additional glusterfs cluster on infra nodes with
Ife73d7c50c304cff7cd05e08f74855cb107f3c46 but never actually enabled
the registry to make use of it. This is controlled by the
openshift_hosted_registry_storage_kind variable which needs to be set
to 'glusterfs'.

Change-Id: I67ae122447fdf0ae32f576b849dfa65cce4f30a3
Closes-Bug: #1804154