Remove validation part of the online database migration pre-upgrade.
That task included a validation in the form of "if the container is
not there fails". This was done to ensure that the online database
migration was run even in the case where the upgrade was run on a
environment where some container would be stopped for some reason.

This validation proves to be problematic as having the related host
services stopped and the container non-running is a "legitimate" state
during the re-run of a fail upgrade.

That validation then completely blocks the upgrade.

We remove the validation part of that tasks as it's very unlikely,
belongs to a validation tasks to be done outside of the upgrade and
block a valid path from working.

Change-Id: I6ca70cb913f7cdd6fc4fbcc70698992e2074dc9c
Closes-Bug: #1804459