Verify docker as part of the online database migration pre-upgrade.
We have to check for docker presence here because it might be missing
during the re-run of a failed upgrade.

This is a sort of backport of
I6ca70cb913f7cdd6fc4fbcc70698992e2074dc9c in the sense that it solves
the same issue.  The patch is so different though that this is a rocky
only patch.

Change-Id: I2e0cebb09b9cefb08a44b64a24fc4159d01d27bc
Closes-Bug: #1804459