ctlplane pre-alloc IPs - deprecated/non-deprecated role name
By default, Compute role template set's the deprecated_param_ips
parameter in roles data. This forces the use of the deprecated
names in paramer_defaults when using predictable IPs for the
ctlplane network.

To allow the user to either use the deprecated role name, or the
non deprecated role name in parameters defaults extend the
ctlplane_fixed_ip_set contition to use or logic to test for data
in either the deprecated name parameter or the new parameter.

In the server resource use yaql to pick the first element that
is not empty. The non-deprecated parameter name is prioritiezed.

Change-Id: Iedc65064c5efaa618c3d54df10bf09296829efd2
Closes-Bug: #1805482
(cherry picked from commit 7a1cd822f648384b24a85b81128c841b59e8d6aa)