Mount puppet-generated directory for tripleo-validations
Tripleo validations relies on the /var/lib/config-data/puppet-generated tree
to check out configuration files from the mistral_executor container. In the
inventory, the undercloud is known as 'localhost' and the validations are trying
to read those configuration files in the mistral_executor file system instead of
the undercloud host.

Change-Id: Ibc034a9fe7cef402743657058c551da00aaa6dc0
Signed-off-by: Gael Chamoulaud <gchamoul@redhat.com>