Mount config-data/puppet-generated/nova for nova_api_ensure_default_cell
When an update changes NovaPassword, we need to run
nova_api_ensure_default_cell container for it to be able to change
the db_connection for the cells in nova_api db. Otherwise
nova_api_discover_hosts which runs all the time fails with db erorr
as the connection string in the database would not change.

Currently, we don't mount either /var/lib/config-data/nova or
/var/lib/config-data/puppet-generated/nova, hence the
TRIPLEO_CONFIG_HASH is not generated for the container and it
does not run during update and may be not in upgrade either.

Change-Id: I0a972796e45a8df614619c95e9d9be9af183b4e5
Closes-Bug: #1805803