Use InternalApi network for openshift_master_cluster_hostname
According to the inventory examples[1] openshift_master_cluster_hostname
points to an internal hostname/address set on the loadbalancer while
openshift_master_cluster_public_hostname points to the external.
This change sets openshift_master_cluster_hostname to use the InternalApi
network instead of the External network as it is at this moment.

[1] https://docs.openshift.com/container-platform/3.11/install/example_inventories.html

Change-Id: I9efab5b07682efd6b03da433801d636e7d324619