Rely on osa defaults for enabled services
Tripleo should rely on openshift-ansible defaults as much as possible,
especially concerning the services which are enabled.

This removes the setting that explicitly disabled the service catalog
and template service broker, and wire in the ansible-service-broker
from heat parameters. The service-catalog and template-service-broker
images do not need to be specified since they're derived from the
oreg_url in openshift-ansible.

Change-Id: Ie09696a54ba01f23003366a9f221d2f32f4b0270
Depends-On: Ic358ed27d63015d297f50c10f553fc1c470ea6a5
Closes-Bug: #1806353