nova_compute fails to start in tls-everywhere configuration
With tls-everywhere enabled connecting to keystone endpoint fails
to retrieve the URL for the placement endpoint as the certificate
can not be verified. While verification is disabled to check the
placement endpoint later, it is not to communicate with keystone.
This disables certificate verification for communication with
keystone.

Related-Bug: 1784155

Change-Id: I317dd62f3a555f375d540a63c21a6fb38d37ca96
(cherry picked from commit a99820a80cd99690c096553dcdc5e7f7472c02f1)