Disable pacemaker on multinode-containers (fs010)
The integration between podman and pacemaker is still work in progress.
While it's being done, let's remove pacemaker from this job, so we can
have at least one CI job running with podman on the overcloud.
We still have coverage for pacemaker in the container-upgrade jobs.

This patch will likely be reverted once we have pacemaker + podman
ready for centos7.

Change-Id: I8638386c8f394c98b047fb2ae4ff97daeeaf4475