puppet_config for rabbitmq_bundle needs file_line
The main reason for this change is that in a related puppet-tripleo patch
we will use the file_line puppet class in order to set some parameters
to reduce the excessive rabbitmq logging.

Because this change will be very specific to rabbitmq running in a
container we also need to make sure that we actually use the
rabbitmq_bundle puppet manifest to create config files and not the
more generic one.

Change-Id: Ic0789da4645a4ee186d82ad7d943de78d4d5c443
Related-Bug: #1806451
(cherry picked from commit 34aa23814ed80be7202ae8d92bdae08ad640f861)