Fix - ApacheServiceBase needs ServiceData
The puppet apache service uses the cidr map in ServiceData.
Services did not pass the ServiceData to the apache
service template. Because of this the property resolves to
an empty string which is not correct. The empty string
cause problems when yaql in common/services.yaml is merging
config_settings.

Closes-Bug: #1806718
Change-Id: Ia3af9535e3af1dad4ac833983ebe29b6002f0815