Spliting compact services in multiples lines
When we generate metadata for compact services, all of the compact
services where joining in one single blob and we hit the size limit
for it. Splitting compact services each one with per line with the
compact service name and a list of their networks, so we will not hit
this size limit anymore.

Change-Id: I90d7bc73000352ad2c4b3f08f00d2a115f075a5e
Depends-On: Ida39f5768c67f982b2fe316f6fae4988a74c8534