Remove console as opendaylight log mechanism
The opendaylight log mechanism is set to "console" in stable/queens
for docker and baremetal templates. This ends up sending logs to the
console. Additionally, the karaf log file size will not be set, and
ends up with the default size of 16MB (vs 2GB), so log files rotate
out very fast on scaled systems.

Closes-Bug: #1807746

https://bugs.launchpad.net/tripleo/+bug/1807746

Change-Id: Ia70949f8cad9aa5d66e18e4a1175297b206b9ab7
Signed-off-by: Victor Pickard <vpickard@redhat.com>