Run local registry and prep cont at standalone
The standalone job were not running yum update on the containers, to do
so we need to specify the updater paremters in the
container-prepare-parameters [1] and also we have to activate the docker
local registry and call the conatiner prepare service.

[1] https://review.openstack.org/#/c/621517/

NOTE: this change differs from master as we do not have a podman service
in Rocky.
Related-Bug: #1808591

Change-Id: I74e817bc9b9dd522db3da7753c91a3884d99f8c8
Related-Bug: #1805968
(cherry picked from commit 62b54268af4c341288e745c356f28124f61e9d2b)