Fix access to /var/lib/haproxy when SELinux is enabled
Currently we don't use relabeling of the folder when SELinux is enabled.
This leads to the fact that we can not update the configuration of
haproxy during the update, because of missing permissions.

This commit adds the relabeling for the folder, which allows the
container with haproxy to write into it.

Closes-Bug: #1807933

Change-Id: Ie79aed5f5665658ea09e000a4847062e9207e25c