Force rhsm re-registration during upgrades.
The redhat-subscription ansible module doesn't cover the case
when a system is already registered via Satellite and we try to
upgrade it's repositories to a new version providing a different
activation key. The role will catch that the system is already
registered and it won't register the new activation key.
This patch passes the rhsm_force_registration option when invoking
redhat-subscription during an upgrade. This way the system will
be unregistered and registered again, activating the new key passed.

Change-Id: I9cd35882e5db47d22df8a456749188c17b48e451
Closes-Bug: #1807987