run docker_puppet_tasks on any role
Currently this assumes all tasks will run on the primary controller
but because of composable roles, that may not be the case.

An example is if you deploy keystone on any role other than the
role tagged primary e.g Controller by default, we don't create
any of the users/endpoints because the tasks aren't written to
the role unless keystone actually runs there.

Closes-Bug: #1792613
Change-Id: Ib6efd03584c95ed4ab997f614aa3178b01877b8c