Gracefully shutdown Mysql before upgrade.
When upgrading from MySQL 10.1 to 10.3 a bug appears if no
shutdown is being performed, as the redo log format has
changed in version 10.3.2 [0].

Make sure we always stop the MySQL server cleanly before
upgrading to a new version, to avoid redo log issue.

Note: to be idempotent, we need to stop the mysql container
rather than delete it; to be able to stop the container, we
amend the restart policy of the mysql container.

[0] - https://jira.mariadb.org/browse/MDEV-14848

Change-Id: Ia07b7755867858c74c7334424e8e6579ace495db
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Closes-Bug: #1810136