Bind mount docker-puppet.py in RO without SElinux labelling
docker-puppet.sh doesn't need to be bind-mounted in read-write,
read-only should be enough.
Therefore, we don't need to relabel the script when running the
container.

The background of this patch can be found here:
https://github.com/containers/libpod/issues/1844

The version of runc that is vendored in libpod contains is a bit old and
doesn't the labelling task isn't tied to the threads yet (done by
https://github.com/opencontainers/runc/commit/aa3fee6c80f2817d47e912372267c74a96cde2bd)

We will request an update of runc in libpod but we also want to avoid
useless RW for this bind mount, which is the goal of this patch.

Note: we also switch /etc/config.pp and /etc/puppet/ to RO, without labelling
as well.

Change-Id: I87568372e80bd8bdb17ae6396ffe5805e37359a7