Reno only - Check for available networks for a role
Since change: I07822ec0cba7eed352c0010eb893b5e5a522e95c
resources are no longer created for networks that are not
defined in roles data. While this is an improvement we
need to communicate this change.

There is tribal knowledge and documentation that failed
reflect the requirement to add networks to roles data
since the introduction of composable networks in Pike.

Prior to Pike adding a network to a role was achived by
overriding the resource_registry entry to not use the
noop.yaml fake port (fall back to ctlplane) template.
i.e to add External network to compute role the
following was commonly added to network-environment.yaml

  OS::TripleO::Compute::Ports::ExternalPort:
    ../network/ports/external.yaml

NOTE: Current OVN-DVR and ODL doc's downstream uses the
      resource_registry override, whitout also adding
      the network to roles data.

Related-Bug: #1800811
Change-Id: I6c03c7a2bd6f369bf35a9e479a97302c9a455197