Run 'Delete Upgrade Flag and Unset it via Rest' only once
Currently the curl commands associated with this command run on all
controller nodes of the overcloud. Becuase the ODL URI is always the
same it's actually doing it against the same REST API every time,
causing problems with the optimistic locking ODL uses.

This patch adds extra error information and limits the execution of this
task to only once per playbook (i.e. just one controller).

Change-Id: I75aed2c0f412961c1eed2ff14e039a0baca09e8a