Fix deployment of gluster with openshift AllInOne
When TripleO deploys openshift with CNS, it configures it to setup
gluster for both the worker nodes and the image registry. However, the
code was incorrectly adding the glusterfs_registry group only for the
Infra node and was forgetting about the AllInOne role.

Closes-Bug: #1811334
Change-Id: Ib502a28db4d67434686a058562a37d1e1df3e263
(cherry picked from commit 87358befc0ab9ce225b5607c6069ad0973e53482)