Create /run/netns if does not exist
Unlike docker with podman you can't mount a host directory
that does not exist. We did some work as part of bug 1799638
take care of this for reboots. However, we need it to be created
for fresh installations, as nodes are not necessarily rebooted
after installing paunch. The real issue without reboots did not
surface before the next patch in this series, as we were using
cni0 bridge network for some containers and that resulted in
namespace being created and the directory existing for
neutron_dhcp and neutron_l3 containers in step 4 to mount them.

This patch creates/removes a temp namespace in host_prep_tasks
to ensure that the directory exists.

Change-Id: I19d660168c98887a5e352b3413235888c800760d
Related-Bug: #1799638
(cherry picked from commit 9b284e74cbdcf43253edaf424534b71a483fdb69)